Complete 06_repeatString
I unfortunately had to glance at the solution
to understand that my syntax with the return
was causing my confusion early on. I have to
keep in mind that in solutions such as this
the return comes after the loop and not within
it. That way the loop does its entire cycle
before returning the variable.